+++ Postal/IMAPSession.swift
@@ -34,7 +34,7 @@ final class IMAPSession {
     let imap: UnsafeMutablePointer<mailimap>
     
     private(set) var capabilities: IMAPCapability = []
-    private var defaultNamespace: IMAPNamespace? = nil
+    private(set) var defaultNamespace: IMAPNamespace? = nil
     private var serverIdentity = IMAPIdentity([:])
     
     private var selectedFolder: String = ""
@@ -199,20 +199,6 @@ final class IMAPSession {
         }
     }
     
-    func listFolders() throws -> [Folder] {
-        let prefix = defaultNamespace?.items.first?.prefix ?? MAIL_DIR_SEPARATOR_S
-        var list: UnsafeMutablePointer<clist> = nil;
-        if capabilities.contains(.XList) && !capabilities.contains(.Gmail) {
-            // XLIST support is deprecated on Gmail. See https://developers.google.com/gmail/imap_extensions#xlist_is_deprecated
-            try mailimap_xlist(imap, prefix, "*", &list).toIMAPError?.check()
-        } else {
-            try mailimap_list(imap, prefix, "*", &list).toIMAPError?.check()
-        }
-        defer { mailimap_list_result_free(list) }
-        
-        return makeFolders(sequence(list, of: mailimap_mailbox_list.self))
-    }
-    
     func select(folder: String) throws -> IMAPFolderInfo {
         if folder != selectedFolder {
             try mailimap_select(imap, folder).toIMAPError?.check()
@@ -223,15 +209,6 @@ final class IMAPSession {
         // store last selected folder
         selectedFolder = folder
         return IMAPFolderInfo(selectionInfo: info)
-    }
-    
-    private func makeFolders<S: SequenceType where S.Generator.Element == mailimap_mailbox_list>(sequence: S) -> [Folder] {
-        return sequence.flatMap { (folder: mailimap_mailbox_list) -> Folder? in
-            guard let name = String.fromCString(folder.mb_name) else { return nil }
-            var mb_delimiter: [CChar] = [ folder.mb_delimiter, 0 ]
-            guard let delimiter = String.fromCString(&mb_delimiter) else { return nil }
-            return Folder(name: name, flags: FolderFlag(flags: folder.mb_flag), delimiter: delimiter)
-        }
     }
     
     func configureIdentity() throws {
